Cleaned the code. Added Swagger configuration. Removed Application configuration
