+++ CitySearch-core/src/main/java/com/app/city/dao/CityDaoInterface.java
@@ -1,6 +1,6 @@
-package com.app.dao;
+package com.app.city.dao;
 
-import com.app.entities.CityInfo;
+import com.app.city.entities.CityInfo;
 
 public interface CityDaoInterface {
   public void addCity(CityInfo city);
