+++ CitySearch-core/src/main/java/com/app/city/services/CityServiceInterface.java
@@ -1,0 +1,7 @@
+package com.app.city.services;
+
+import com.app.city.models.City;
+
+public interface CityServiceInterface {
+	public void processAddCity(City city);
+}
